+++ tfe/resource_tfe_policy_set.go
@@ -102,6 +102,7 @@ func resourceTFEPolicySet() *schema.Resource {
 			"workspace_ids": {
 				Type:          schema.TypeSet,
 				Optional:      true,
+				Computed:      true,
 				Elem:          &schema.Schema{Type: schema.TypeString},
 				ConflictsWith: []string{"global"},
 			},